+++ packages/ui/src/components/index.ts
@@ -35,8 +35,10 @@ export {
   MobileStepper,
   FormControl,
   Slide,
+  Grid,
   InputAdornment,
   Snackbar,
+  Radio,
 } from '@mui/material';
 
 export type {
@@ -76,6 +78,7 @@ export type {
   FormControlProps,
   InputAdornmentProps,
   SlideProps,
+  GridProps,
   SnackbarProps,
 } from '@mui/material';
 
@@ -94,3 +97,4 @@ export * from './CopyButton';
 export * from './OtpInput';
 export * from './Typography';
 export * from './Banner';
+export * from './AddressDropdown';