+++ src/components/ProductTours/WalletProductTour.tsx
@@ -28,13 +28,14 @@ export const WalletProductTour = ({ onClose }: WalletProductTourProps) => {
   const handleCloseIntroduceStartDialog = useCallback(() => {
     setIntroduceStartDialog(false);
     localStorage.setItem('showProductTour', 'false');
-    localStorage.setItem('showProductTourSnackbar', 'true');
+    localStorage.setItem('showProductTourSnackbar', 'false');
     onClose?.();
   }, [setIntroduceStartDialog, onClose]);
 
   const handleCloseIntroduceFinishDialog = useCallback(() => {
     setIntroduceFinishDialog(false);
     localStorage.setItem('showProductTour', 'false');
+    localStorage.setItem('showProductTourSnackbar', 'true');
     onClose?.();
   }, [setIntroduceFinishDialog, onClose]);